requirements.txt
flask
flask_restful
flask_httpauth
pyjwt
jsonmerge
json